chore: refactor protocol client generator
